+++ packages/liquid-html-parser/src/stage-2-ast.ts
@@ -71,6 +71,7 @@ import {
   ConcreteHtmlRawTag,
   ConcreteLiquidRawTag,
   LiquidHtmlConcreteNode,
+  ConcreteLiquidTagBaseCase,
 } from './stage-1-cst';
 import { Comparators, NamedTags, NodeTypes, nonTraversableProperties, Position } from './types';
 import { assertNever, deepGet, dropLast } from './utils';
@@ -552,7 +553,6 @@ export interface LiquidVariableLookup extends ASTNode<NodeTypes.VariableLookup> 
 export type HtmlNode =
   | HtmlComment
   | HtmlElement
-  | HtmlDanglingMarkerOpen
   | HtmlDanglingMarkerClose
   | HtmlVoidElement
   | HtmlSelfClosingElement
@@ -571,25 +571,6 @@ export interface HtmlElement extends HtmlNodeBase<NodeTypes.HtmlElement> {
 
   /** The range covered by the end tag */
   blockEndPosition: Position;
-}
-
-/**
- * The node used to represent opening tags without its matching close tag, does not have children nor a close tag.
- *
- * Typically found inside if statements.
- *
- * ```
- * {% if cond %}
- *   <wrapper>
- * {% endif %}
- * ```
- */
-export interface HtmlDanglingMarkerOpen extends HtmlNodeBase<NodeTypes.HtmlDanglingMarkerOpen> {
-  /**
-   * The name of the tag can be compound
-   * e.g. `<{{ header_type }}--header />`
-   */
-  name: (TextNode | LiquidVariableOutput)[];
 }
 
 /**
@@ -776,9 +757,7 @@ export interface ASTNode<T> {
 }
 
 interface ASTBuildOptions {
-  /**
-   * Whether the parser should throw if the document node isn't closed
-   */
+  /** Whether the parser should throw if the document node isn't closed */
   allowUnclosedDocumentNode: boolean;
 
   /**
@@ -795,15 +774,10 @@ export function isBranchedTag(node: LiquidHtmlNode) {
   return node.type === NodeTypes.LiquidTag && ['if', 'for', 'unless', 'case'].includes(node.name);
 }
 
-// Not exported because you can use node.type === NodeTypes.LiquidBranch.
-function isLiquidBranchDisguisedAsTag(node: LiquidHtmlNode): node is LiquidTagBaseCase {
-  return node.type === NodeTypes.LiquidTag && ['else', 'elsif', 'when'].includes(node.name);
-}
-
 function isConcreteLiquidBranchDisguisedAsTag(
   node: LiquidHtmlConcreteNode,
-): node is ConcreteLiquidNode & { name: 'else' | 'eslif' | 'when' } {
-  return node.type === ConcreteNodeTypes.LiquidTag && ['else', 'eslif', 'when'].includes(node.name);
+): node is ConcreteLiquidTag & { name: 'else' | 'elsif' | 'when' } {
+  return node.type === ConcreteNodeTypes.LiquidTag && ['else', 'elsif', 'when'].includes(node.name);
 }
 
 export function toLiquidAST(
@@ -850,7 +824,7 @@ export function toLiquidHtmlAST(
   return root;
 }
 
-type LiquidCSTNode = LiquidCST[number];
+type ConcreteCloseNode = ConcreteLiquidTagClose | ConcreteHtmlTagClose;
 
 class ASTBuilder {
   /** the AST is what we're building incrementally */
@@ -882,22 +856,24 @@ class ASTBuilder {
     this.source = source;
   }
 
+  // Returns the array to push nodes to.
   get current() {
     return deepGet<LiquidHtmlNode[]>(this.cursor, this.ast) as LiquidHtmlNode[];
   }
 
+  // Returns the position of the current node in the array
   get currentPosition(): number {
     return (this.current || []).length - 1;
   }
 
   get parent(): ParentNode | undefined {
     if (this.cursor.length == 0) return undefined;
-    return deepGet<LiquidTag | HtmlElement>(dropLast(1, this.cursor), this.ast);
+    return deepGet<ParentNode>(dropLast(1, this.cursor), this.ast);
   }
 
   get grandparent(): ParentNode | undefined {
     if (this.cursor.length < 4) return undefined;
-    return deepGet<LiquidTag | HtmlElement>(dropLast(3, this.cursor), this.ast);
+    return deepGet<ParentNode>(dropLast(3, this.cursor), this.ast);
   }
 
   open(node: LiquidHtmlNode) {
@@ -911,23 +887,31 @@ class ASTBuilder {
   }
 
   push(node: LiquidHtmlNode) {
-    if (node.type === NodeTypes.LiquidTag && isLiquidBranchDisguisedAsTag(node)) {
-      this.closeParentWith(node);
-      this.open(toNamedLiquidBranchBaseCase(node));
-    } else if (node.type === NodeTypes.LiquidBranch) {
-      this.closeParentWith(node);
+    if (node.type === NodeTypes.LiquidBranch) {
+      const previousBranch = this.findCloseableParentBranch(node);
+      if (previousBranch) {
+        // close dangling open HTML nodes
+        while (
+          this.parent &&
+          (this.parent as ParentNode) !== previousBranch &&
+          this.parent.type === NodeTypes.HtmlElement
+        ) {
+          // 0-length blockEndPosition at the position of the next branch
+          this.parent.blockEndPosition = { start: node.position.start, end: node.position.start };
+          this.closeParentWith(node);
+        }
+        // close the previous branch
+        this.closeParentWith(node);
+      }
       this.open(node);
     } else {
       this.current.push(node);
     }
   }
 
-  close(
-    node: ConcreteLiquidTagClose | ConcreteHtmlTagClose,
-    nodeType: NodeTypes.LiquidTag | NodeTypes.HtmlElement,
-  ) {
+  close(node: ConcreteCloseNode, nodeType: NodeTypes.LiquidTag | NodeTypes.HtmlElement) {
     if (isLiquidBranch(this.parent)) {
-      this.closeParentWith(node as LiquidCSTNode);
+      this.closeParentWith(node);
     }
 
     if (!this.parent) {
@@ -940,14 +924,25 @@ class ASTBuilder {
     }
 
     if (getName(this.parent) !== getName(node) || this.parent.type !== nodeType) {
-      throw new LiquidHTMLASTParsingError(
-        `Attempting to close ${nodeType} '${getName(node)}' before ${this.parent.type} '${getName(
-          this.parent,
-        )}' was closed`,
-        this.source,
-        this.parent.position.start,
-        node.locEnd,
-      );
+      const suitableParent = this.findCloseableParentNode(node);
+
+      if (this.parent.type === NodeTypes.HtmlElement && suitableParent) {
+        // close dangling open HTML nodes
+        while ((this.parent as ParentNode) !== suitableParent) {
+          // 0-length end block position
+          this.parent.blockEndPosition = { start: node.locStart, end: node.locStart };
+          this.closeParentWith(node);
+        }
+      } else {
+        throw new LiquidHTMLASTParsingError(
+          `Attempting to close ${nodeType} '${getName(node)}' before ${this.parent.type} '${getName(
+            this.parent,
+          )}' was closed`,
+          this.source,
+          this.parent.position.start,
+          node.locEnd,
+        );
+      }
     }
 
     // The parent end is the end of the outer tag.
@@ -961,8 +956,55 @@ class ASTBuilder {
     this.cursor.pop();
   }
 
-  closeParentWith(next: LiquidHtmlNode | LiquidCSTNode) {
-    if (isLiquidBranch(this.parent)) {
+  // This function performs the following tasks:
+  // - Tries to find a parent branch to close when pushing a new branch.
+  // - This is necessary because we allow unclosed HTML element nodes.
+  // - The function traverses up the tree until it finds a LiquidBranch.
+  // - If it encounters anything other than an Unclosed HTML Element, it throws.
+  findCloseableParentBranch(next: LiquidBranch): LiquidBranch | null {
+    for (let index = this.cursor.length - 1; index > 0; index -= 2) {
+      const parent = deepGet<ParentNode>(this.cursor.slice(0, index), this.ast);
+      const parentProperty = this.cursor[index] as string;
+      const isUnclosedHtmlElement =
+        parent.type === NodeTypes.HtmlElement && parentProperty === 'children';
+      if (parent.type === NodeTypes.LiquidBranch) {
+        return parent;
+      } else if (!isUnclosedHtmlElement) {
+        throw new LiquidHTMLASTParsingError(
+          `Attempting to open LiquidBranch '${next.name}' before ${parent.type} '${getName(
+            parent,
+          )}' was closed`,
+          this.source,
+          parent.position.start,
+          next.position.end,
+        );
+      }
+    }
+    return null;
+  }
+
+  // Check if there's a parent in the ancestry that this node correctly closes
+  findCloseableParentNode(
+    current: ConcreteHtmlTagClose | ConcreteLiquidTagClose,
+  ): LiquidTag | null {
+    for (let index = this.cursor.length - 1; index > 0; index -= 2) {
+      const parent = deepGet<ParentNode>(this.cursor.slice(0, index), this.ast);
+      if (
+        getName(parent) === getName(current) &&
+        parent.type === NodeTypes.LiquidTag &&
+        ['if', 'unless', 'case'].includes(parent.name)
+      ) {
+        return parent;
+      } else if (parent.type === NodeTypes.LiquidTag) {
+        return null;
+      }
+    }
+    return null;
+  }
+
+  // sets the parent's end position to the start of the next one.
+  closeParentWith(next: LiquidHtmlNode | ConcreteCloseNode) {
+    if (this.parent) {
       if ('locStart' in next) {
         this.parent.position.end = next.locStart;
       } else {
@@ -984,7 +1026,6 @@ function getName(
   if (!node) return null;
   switch (node.type) {
     case NodeTypes.HtmlElement:
-    case NodeTypes.HtmlDanglingMarkerOpen:
     case NodeTypes.HtmlDanglingMarkerClose:
     case NodeTypes.HtmlSelfClosingElement:
     case ConcreteNodeTypes.HtmlTagClose:
@@ -1100,11 +1141,7 @@ function buildAst(
       }
 
       case ConcreteNodeTypes.HtmlTagOpen: {
-        if (isAcceptableDanglingMarkerOpen(builder, cst as LiquidHtmlCST, i)) {
-          builder.push(toHtmlDanglingMarkerOpen(node, options));
-        } else {
-          builder.open(toHtmlElement(node, options));
-        }
+        builder.open(toHtmlElement(node, options));
         break;
       }
 
@@ -1299,6 +1336,9 @@ function toLiquidTag(
 ): LiquidTag | LiquidBranch {
   if (typeof node.markup !== 'string') {
     return toNamedLiquidTag(node as ConcreteLiquidTagNamed, options);
+  } else if (isConcreteLiquidBranchDisguisedAsTag(node)) {
+    // `elsif`, `else`, `case`, but with unparseable markup.
+    return toNamedLiquidBranchBaseCase(node);
   } else if (options.isBlockTag) {
     return {
       name: node.name,
@@ -1463,16 +1503,16 @@ function toNamedLiquidTag(
   }
 }
 
-function toNamedLiquidBranchBaseCase(node: LiquidTagBaseCase): LiquidBranchBaseCase {
+function toNamedLiquidBranchBaseCase(node: ConcreteLiquidTagBaseCase): LiquidBranchBaseCase {
   return {
     name: node.name,
     type: NodeTypes.LiquidBranch,
-    markup: node.markup,
-    position: { ...node.position },
+    markup: node.name !== 'else' ? node.markup : '', // stripping superfluous else stuff...
+    position: { start: node.locStart, end: node.locEnd },
     children: [],
-    blockStartPosition: { ...node.position },
-    whitespaceStart: node.whitespaceStart,
-    whitespaceEnd: node.whitespaceEnd,
+    blockStartPosition: { start: node.locStart, end: node.locEnd },
+    whitespaceStart: node.whitespaceStart ?? '',
+    whitespaceEnd: node.whitespaceEnd ?? '',
     source: node.source,
   };
 }
@@ -1826,20 +1866,6 @@ function toHtmlElement(node: ConcreteHtmlTagOpen, options: ASTBuildOptions): Htm
   };
 }
 
-function toHtmlDanglingMarkerOpen(
-  node: ConcreteHtmlTagOpen,
-  options: ASTBuildOptions,
-): HtmlDanglingMarkerOpen {
-  return {
-    type: NodeTypes.HtmlDanglingMarkerOpen,
-    name: cstToAst(node.name, options) as (TextNode | LiquidVariableOutput)[],
-    attributes: toAttributes(node.attrList || [], options),
-    position: position(node),
-    blockStartPosition: position(node),
-    source: node.source,
-  };
-}
-
 function toHtmlDanglingMarkerClose(
   node: ConcreteHtmlTagClose,
   options: ASTBuildOptions,
@@ -1892,14 +1918,6 @@ function toTextNode(node: ConcreteTextNode): TextNode {
 
 const MAX_NUMBER_OF_SIBLING_DANGLING_NODES = 2;
 
-function isAcceptableDanglingMarkerOpen(
-  builder: ASTBuilder,
-  cst: LiquidHtmlCST,
-  currIndex: number,
-): boolean {
-  return isAcceptableDanglingMarker(builder, cst, currIndex, ConcreteNodeTypes.HtmlTagOpen);
-}
-
 function isAcceptableDanglingMarkerClose(
   builder: ASTBuilder,
   cst: LiquidHtmlCST,
@@ -1915,7 +1933,7 @@ function isAcceptableDanglingMarkerClose(
     return !parentIsOfCorrectName;
   }
 
-  return isAcceptableDanglingMarker(builder, cst, currIndex, ConcreteNodeTypes.HtmlTagClose);
+  return isAcceptableDanglingMarker(builder);
 }
 
 // This function checks that the builder.current node accepts dangling nodes.
@@ -1923,59 +1941,13 @@ function isAcceptableDanglingMarkerClose(
 // The current logic is:
 //  - Grandparent node must be an if-like statement
 //  - Parent node must be a LiquidBranch
-//  - All sibling nodes must be of the same type (all close or all open)
-//
-// I want to extend that to pushing dangling close in completion mode...
-//
-// So that I can complete </[cursor]> and not have a broken AST.
-//
-// That would mean that the partial ast looks like this
-// <a>
-//   ...children
-//   </[cursor]>
-function isAcceptableDanglingMarker(
-  builder: ASTBuilder,
-  cst: LiquidHtmlCST,
-  currIndex: number,
-  nodeType: ConcreteNodeTypes.HtmlTagOpen | ConcreteNodeTypes.HtmlTagClose,
-): boolean {
-  if (!isAcceptingDanglingMarkers(builder, nodeType)) {
-    return false;
-  }
-
-  const maxIndex = Math.min(
-    cst.length,
-    currIndex + MAX_NUMBER_OF_SIBLING_DANGLING_NODES - builder.current.length,
-  );
-
-  for (let i = currIndex; i <= maxIndex; i++) {
-    if (isConcreteExceptionEnd(cst[i])) {
-      return true;
-    }
-    if (cst[i].type !== nodeType) {
-      return false;
-    }
-  }
-
-  return false;
-}
-
-const DanglingMapping = {
-  [ConcreteNodeTypes.HtmlTagOpen]: NodeTypes.HtmlDanglingMarkerOpen,
-  [ConcreteNodeTypes.HtmlTagClose]: NodeTypes.HtmlDanglingMarkerClose,
-} as const;
-
-function isAcceptingDanglingMarkers(
-  builder: ASTBuilder,
-  nodeType: ConcreteNodeTypes.HtmlTagOpen | ConcreteNodeTypes.HtmlTagClose,
-) {
+function isAcceptableDanglingMarker(builder: ASTBuilder): boolean {
   const { parent, grandparent } = builder;
   if (!parent || !grandparent) return false;
   return (
     parent.type === NodeTypes.LiquidBranch &&
     grandparent.type === NodeTypes.LiquidTag &&
-    ['if', 'unless', 'case'].includes(grandparent.name) &&
-    builder.current.every((node) => node.type === DanglingMapping[nodeType])
+    ['if', 'unless', 'case'].includes(grandparent.name)
   );
 }