+++ packages/liquid-html-parser/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @shopify/liquid-html-parser
 
+## 2.0.0
+
+### Major Changes
+
+- 636895f: Add support for unclosed HTML Element nodes inside branching code
+
+  **Breaking changes**
+
+  - `HtmlDanglingMarkerOpen` nodes are folded into `HtmlElement` nodes
+
+    - You can identify a node without its closing tag by looking at `blockEndPosition`
+
+      ```js
+      node.source.slice(node.blockEndPosition.start, node.blockEndPosition.end) === '';
+      ```
+
+  **Why the change**?
+
+  - `HtmlDanglingMarkerOpen` nodes did not have children, since we added support for nodes without closing markers, we removed that node type entirely.
+
+### Minor Changes
+
+- 636895f: Allow 'HtmlDanglingMarkerClose' nodes to have non-dangling marker siblings
+
 ## 1.1.1
 
 ### Patch Changes