+++ packages/vscode-extension/CHANGELOG.md
@@ -1,5 +1,54 @@
 ## theme-check-vscode
 
+## 1.15.0
+
+### Minor Changes
+
+- a780181: Improve support of unclosed HTML nodes inside conditional code
+
+  The following code is now supported by the Liquid prettier plugin and the `"shopifyLiquid.themeCheckNextDevPreview"` VS Code setting.
+
+  ```liquid
+  <div>
+    {% if is_accordion %}
+      <details>
+        <summary>
+    {% else %}
+      <h3>
+    {% endif %}
+
+    ... summary/header content ...
+
+    {% if is_accordion %}
+      </summary>
+    {% else %}
+      <a href='{{ bookmark }}'>
+        <svg>
+          ...
+        </svg>
+      </a>
+      </h3>
+    {% endif %}
+
+    ... body content ...
+
+    {% if is_accordion %}
+      </details>
+    {% endif %}
+  </div>
+  ```
+
+  Code that doesn’t properly close across branches will be flagged by a new theme check: `UnclosedHTMLElement`. See [PR](https://github.com/Shopify/theme-tools/pull/254) for details.
+
+### Patch Changes
+
+- Updated dependencies [636895f]
+- Updated dependencies [636895f]
+- Updated dependencies [636895f]
+  - @shopify/prettier-plugin-liquid@1.4.0
+  - @shopify/liquid-html-parser@2.0.0
+  - @shopify/theme-language-server-node@1.7.1
+
 ## 1.14.0
 
 ### Minor Changes